requirements.txt
psutil==5.9.5
litellm==1.66.3
deprecated
google-cloud-aiplatform
google-cloud-storage==2.19.0    
google-genai==1.14.0
gitpython==3.1.36
google-adk==0.5.0
python-dotenv==1.0.1
ipykernel
flask
pyyaml
pathlib
streamlit 
flask-cors
flask
llama-index>=0.10.0
llama-index-embeddings-huggingface
llama-index-vector-stores-chroma
chromadb